olpc-configure: unmute master on boot
During testing of a0fb722 it was found that muting master using
alsamixer would be preserved over a reboot.  No known activity mutes the
master control, and the Sugar Frame mute button is implemented by
lowering the volume to zero, but it remains a possibility.

So we unmute it too.